Windows fix (smaller type conversion)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@9960 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation2/components/CCmpTerritoryManager.cpp
@@ -300,7 +300,7 @@ void CCmpTerritoryManager::CalculateTerritories()
 	{
 		for (u32 i = 0; i < tilesW; ++i)
 		{
-			u8 g = passGrid.get(i, j);
+			u16 g = passGrid.get(i, j);
 			u8 cost;
 			if (g & passClassUnrestricted)
 				cost = 255; // off the world; use maximum cost
